Add files via upload
This code is a collection of SQL queries that analyze data related to the COVID-19 pandemic, including case and mortality rates, vaccination data, and global trends. The queries use data from two tables, "CovidDeaths" and "CovidVaccinations," and provide insights on specific countries, continents, and the world as a whole. The code includes temporary tables and a view that can be used for later visualizations.